* Further updates.
I made a few small changes to the display tab choices and added a few more choices on the general tab under specified period and specified last  - figured this was my last opportunity to easily change.  Once I submit it changes will have to be done thru the normal bug reports.  If you don't mind, please test it.  Also I added an optional  imbalance flag to both reports.  The settings default is to turn the imbalance check on.  If the program checked for an imbalance it prints a dash "-" at the bottom of the report.

The new  income-expense-statement  will show up on the report   -  income and  expense   menu as a duplicate income statement   and duplicate profit & loss.  I am hoping they will replace the current versions with the new one.   Thanks