New Report: current-vs-average.scm
This report allows you to pick a time period and then compare it with the
average This report has all of the functionality of  the multiple periods side
by side report, it just adds some additional options.

Place the attached files in the .gnucash folder.  On windows this is
c:\users\YOURID\.gnucash - If you already have a config.user file,
just add the entries in this config.user file to it.