SEPTEMBER 2017
TRANSACTION

I have attached a copy of my modified version of the transaction report.
This version has the following features

1. On the display tab you can select to consolidate transactions and the
report will combine all similar transactions into one total entry.  By
checking and unchecking entries on the display tab you can change how
the numbers are combined - by description, memo, notes, account.  The date
ranged which is combined is controlled based on the date range selected
on the sorting tab.  You can combine for an entire year, by month, by week,
or by quarter.

2. You can use the find option on the sorting tab to limit the report to
only descriptions containing specific text and you can limit the amount
to a given range such as only amounts greater than 500.

3. You can specify on the general tab you want all of the entries to be
scaled.  By which I mean you can take the results for a full year and
divide it by 365 to see what the amounts are on a daily basis.  You can
also take the numbers for a quarter and multiply by 4 to see what they
would look like on an annual basis.

4. You can easily select a time period on the general tab.  Such as last
quarter, 2nd half of the year, or the month of January.  I have also
modified the Income statement in the "other" menu to provide this same
capability.

To install, download the attached files and put them into your .gnucash
folder.  In windows for me this is c:/users/UserName/.gnucash.  Warning
if you already have a config.user file then first put into a different
folder and then copy the entries in config.user into your existing config.user
file.  The transaction  report will show up under reports, the income
statement  in "Other".

---

USER-COLUMN-TRANSACTIONS

This report allows you to control what is printed in each column.  For
example column 1 could be all entries with London in the description and
column 2 could be all entries in account name expenses:travel or just
travel that do not have London in the description.

There are a lot of items in each of the options tab so be sure to scroll
down in each one.  To see what will be displayed you may want to first
run the report with show find  checked in the display tab  and with
descriptions checked further down in the display tab.

Put the three files in your .gnucash folder.  If you have a config.user
then DO NOT OVERWRITE it, put the config.user file in a different folder
and add the line in this config.user to your existing one.

Thanks to those that have tested it so far, and thanks to Milan KnÃ­Å¾ek
for helping to trouble shoot it.

If any one knows how to fix it to work better with style sheets, please
fix and post new files.

If you have problems getting it installed or in getting it to work, let
me know.  Let me know if you have any ideas for improvement.